Add func addVideoToHistory
