+++ YouTube/Classes/Common/Message.swift
@@ -1,0 +1,9 @@
+//
+//  Message.swift
+//  YouTube
+//
+//  Created by Duy Tang on 8/22/16.
+//  Copyright © 2016 Duy Tang. All rights reserved.
+//
+
+import Foundation
